Showing onboarding status on teams page with more refinement.
